Clear up free structures
